# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/books/{id}:
    get:
      tags:
      - api-book-controller
      operationId: getBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    put:
      tags:
      - api-book-controller
      operationId: editBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      - name: name
        in: query
        required: false
        schema:
          type: string
      - name: author
        in: query
        required: false
        schema:
          type: string
      - name: ISBN
        in: query
        required: false
        schema:
          type: string
      - name: pages
        in: query
        required: false
        schema:
          type: string
      - name: genre
        in: query
        required: false
        schema:
          type: string
      - name: date
        in: query
        required: false
        schema:
          type: string
      - name: publisher
        in: query
        required: false
        schema:
          type: string
      - name: series
        in: query
        required: false
        schema:
          type: string
      - name: description
        in: query
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - api-book-controller
      operationId: deleteBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
...
  /api/books/{id}/wanted:
    post:
      tags:
      - api-user-lists-controller
      summary: Add book to wanted list
      operationId: addWantedBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Book added to wanted list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_BasicInfo'
        "401":
          description: Unauthorized
        "400":
          description: Invalid book ID
        "404":
          description: Book not found
  /api/books/{id}/reading:
    post:
      tags:
      - api-user-lists-controller
      summary: Add book to reading list
      operationId: addReadingBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Book added to reading list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_BasicInfo'
        "401":
          description: Unauthorized
        "400":
          description: Invalid book ID
        "404":
          description: Book not found
  /api/books/{id}/read:
    post:
      tags:
      - api-user-lists-controller
      summary: Add book to read list
      operationId: addReadBook
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: Unauthorized
        "200":
          description: Book added to read list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_BasicInfo'
        "400":
          description: Invalid book ID
        "404":
          description: Book not found
...
  /api/users/{username}/books:
    get:
      tags:
      - api-user-lists-controller
      summary: Get a user's specific list
      operationId: getUserLists
      parameters:
      - name: username
        in: path
        required: true
        schema:
          type: string
      - name: list
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: List found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_BasicInfo'
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                type: object
        "400":
          description: Invalid list
          content:
            '*/*':
              schema:
                type: object
...
  /api/users/me:
    get:
      tags:
      - api-user-lists-controller
      summary: Get current user's info
      operationId: getUserInfo
      responses:
        "200":
          description: User logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User_BasicInfo'
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
...
  /api/books/me:
    get:
      tags:
      - api-user-lists-controller
      summary: Get current user's specific list
      operationId: getCurrentUserLists
      parameters:
      - name: list
        in: query
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: List found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_BasicInfo'
        "400":
          description: Invalid list
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Unauthorized
          content:
            '*/*':
              schema:
                type: object
...
  /api/books/{id}/lists:
    delete:
      tags:
      - api-user-lists-controller
      summary: Remove book from all lists
      operationId: removeBookFromLists
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "401":
          description: Unauthorized
        "400":
          description: Invalid book ID
        "200":
          description: Book removed from all lists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Book_BasicInfo'
        "404":
          description: Book not found
components:
  schemas:
...
    Book_BasicInfo:
      type: object
      properties:
        ID:
          type: integer
          format: int64
        title:
          type: string
        authorString:
          type: string
        description:
          type: string
        releaseDate:
          type: string
        ISBN:
          type: string
        averageRating:
          type: number
          format: double
        series:
          type: string
        pageCount:
          type: integer
          format: int32
        publisher:
          type: string
        id:
          type: integer
          format: int64
        isbn:
          type: string
...
    User_BasicInfo:
      type: object
      properties:
        readBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_BasicInfo'
        readingBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_BasicInfo'
        wantedBooks:
          type: array
          items:
            $ref: '#/components/schemas/Book_BasicInfo'
        role:
          type: array
          items:
            type: string
...